Do not request wasapi buffer if frame count is 0
Wasapi buffer pointer was requested even if the requested number of frames was 0, and the code assumed that the pointer
is non-null if the request succeeded. According to Microsoft docs, the given pointer variable is not written to if the
requested frame count is 0. See: https://docs.microsoft.com/en-us/windows/win32/api/audioclient/nf-audioclient-iaudiorenderclient-getbuffer